#353 Factor conversation and rendering into separate component
